+++ package.json
@@ -23,6 +23,7 @@
     "@apollo/client": "^3.7.16",
     "@apollo/server": "^4.7.5",
     "@auth/prisma-adapter": "^1.0.0",
+    "@editorjs/attaches": "^1.3.0",
     "@editorjs/checklist": "^1.5.0",
     "@editorjs/code": "^2.8.0",
     "@editorjs/delimiter": "^1.3.0",
