+++ src/components/Dropzone/UploadProviders/cloudinary.ts
@@ -1,12 +1,10 @@
 import getConfig from 'next/config'
 
-import { CLOUDFLARE_IMAGE_DELIVERY_BASE_URL } from '~/lib/cloudflare'
-
 const { publicRuntimeConfig } = getConfig()
 const { CLOUDINARY_API_KEY, CLOUDINARY_CLOUD_NAME, CLOUDINARY_PRESET } =
   publicRuntimeConfig
 
-export const Cloudinary = {
+const Cloudinary = {
   getParams: (site, extra_params) => {
     const timestamp = Date.now()
     const params = {
@@ -55,56 +53,4 @@ export const Cloudinary = {
   },
 }
 
-export const Cloudflare = {
-  getParams: () => {
-    return {}
-  },
-  sign: async (req) => {
-    return await fetch('/api/images/cloudflare/sign').then((res) => res.json())
-  },
-  uploadFile: async ({ site, params, file, signData }) => {
-    const { uploadUrl } = signData
-    const data = new FormData()
-    data.append('file', file)
-    const upload = await fetch(uploadUrl, {
-      method: 'POST',
-      body: data,
-    }).then((r) => r.json())
-    return upload
-  },
-  getPublicUrl: (upload) => {
-    return `${CLOUDFLARE_IMAGE_DELIVERY_BASE_URL}/${upload?.result?.id}/public`
-  },
-}
-
-export function uploadFile({
-  file,
-  fileProvider,
-  site,
-  upload_options,
-}): Promise<string> {
-  return new Promise(async (resolve, reject) => {
-    const params = fileProvider.getParams(site, upload_options)
-    const timestamp = Date.now()
-    const signData = await fileProvider.sign(params)
-
-    if (!signData) {
-      console.error('No signed url')
-      reject()
-    }
-
-    const upload = await fileProvider.uploadFile({
-      site: site,
-      params,
-      file,
-      signData,
-    })
-    if (!upload) {
-      console.error('Upload failed')
-      reject()
-    }
-
-    const url = fileProvider.getPublicUrl(upload)
-    resolve(url)
-  })
-}
+export default Cloudinary
